+++ src/custom/CGFCardTemplate.gd
@@ -1,0 +1,22 @@
+extends Card
+
+
+func _ready() -> void:
+	pass
+
+# Sample code on how to ensure costs are paid when a card
+# is dragged from hand to the table
+func common_move_scripts(new_container: Node, old_container: Node) -> void:
+	if new_container == cfc.NMAP.board and old_container == cfc.NMAP.hand:
+		pay_play_costs()
+
+
+# Sample code on how to figure out costs of a card
+func get_modified_credits_cost() -> int:
+	var modified_cost : int = properties.get("Cost", 0)
+	return(modified_cost)
+
+
+# This sample function ensures all costs defined by a card are paid.
+func pay_play_costs() -> void:
+	cfc.NMAP.board.counters.mod_counter("credits", -get_modified_credits_cost())